dedup HttpServerSpec
